Convert functioning example to hypertrace agent model phase 0.
