--- instrumentation/instrumentation-python-agent/README.md
@@ -1,20 +1,0 @@
-# Python Agent
-
-## Objectives
-
-
-
-* Created a working example of python opentelemetry agent. 
-* Added  trace feature
-* Added python agent to the startup code.
-* Configured Agent to track HTTP request and responses to / from web services.
-* Created a instrumentation package  for python 
-* Activated  each instrumentation during its initialization.
-
-## Walkthrough
-
-### A simple Python Flask program 
-
-Calls a dummy backend endpoint
-
-
