--- instrumentation/instrumentation-python-agent/python_agent.py
@@ -1,33 +1,0 @@
-# flask_telemetry_agent.py
-# Issue2 - Working example of python opentelemetry agent.
-import flask
-import requests
-
-from opentelemetry import trace
-from opentelemetry.instrumentation.flask import FlaskInstrumentor
-from opentelemetry.instrumentation.requests import RequestsInstrumentor
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import (
-    ConsoleSpanExporter,
-    SimpleExportSpanProcessor,
-)
-
-trace.set_tracer_provider(TracerProvider())
-trace.get_tracer_provider().add_span_processor(
-    SimpleExportSpanProcessor(ConsoleSpanExporter())
-)
-
-app = flask.Flask(__name__)
-FlaskInstrumentor().instrument_app(app)
-RequestsInstrumentor().instrument()
-
-
-@app.route("/")
-def hello():
-    tracer = trace.get_tracer(__name__)
-    with tracer.start_as_current_span("demo-request"):
-        requests.get("https://httpbin.org/anything")
-    return "hello"
-
-
-app.run(debug=True, port=5000)
