+++ test/run_test.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+#File: run_test.sh
+#Author: Nitin Sahai
+#Date: 03/04/2021
+#Notes:
+#
+# This runs the flask test on the demo HyperTrace Python Agent
+#
+curl -X GET http://localhost:5000/
+echo ""
+exit 0
